+++ chat2rag/core/strategies/agent.py
@@ -68,11 +68,10 @@ class AgentStrategy(ResponseStrategy):
                 query=query,
                 top_k=self.request.top_k,
                 score_threshold=self.request.score_threshold,
-                doc_type="qa_pair",
+                filters={"field": "meta.type", "operator": "==", "value": "qa_pair"},
                 messages=history_messages,
                 extra_params=self.request.extra_params | current_time,
                 streaming_callback=self.handler.callback,
-                # generation_kwargs=generation_kwargs,
             )
 
             # 更新聊天历史